+++ Coursework/2D Computer Graphics Assignment/SFML_BulletSpriteObject.cpp
@@ -23,7 +23,7 @@ void SFML_BulletSpriteObject::setMoveDirection(sf::Vector2f directionVector, boo
 
 	if (rotate)
 	{
-		const float PI = 3.1415;
+		const float PI = static_cast<float>(3.1415);
 		float angleRadian = atan2(directionVector.y, directionVector.x);
 		float angleDegree = angleRadian / PI * 180;
 		setRotation(angleDegree);
@@ -59,6 +59,6 @@ void SFML_BulletSpriteObject::applyTexture(sf::Texture* texture)
 	m_initialised = true;
 	m_spriteTexture = texture;
 	// Setting default origin and rectangle
-	m_spriteOrigin = sf::Vector2f(m_spriteTexture->getSize().x / 2, m_spriteTexture->getSize().y / 2);
+	m_spriteOrigin = sf::Vector2f(static_cast<float>(m_spriteTexture->getSize().x / 2), static_cast<float>(m_spriteTexture->getSize().y / 2));
 	m_textureRectangle = sf::IntRect(0, 0, m_spriteTexture->getSize().x, m_spriteTexture->getSize().y);
 }
